+++ shared/hedvig-ui/index.ts
@@ -1,3 +1,19 @@
 import { Badge } from './Badge/badge'
+import {
+  Button,
+  buttonColorMap,
+  ButtonLink,
+  ButtonProps,
+  ButtonsGroup,
+  buttonSizeMap,
+} from './Button/button'
 
-export { Badge }
+export {
+  Badge,
+  Button,
+  ButtonsGroup,
+  buttonSizeMap,
+  buttonColorMap,
+  ButtonLink,
+  ButtonProps,
+}
