+++ src/components/claims/claim-details/components/ClaimNotes.tsx
@@ -9,9 +9,9 @@ import { Spinner } from 'hedvig-ui/sipnner'
 import React, { useState } from 'react'
 
 import styled from '@emotion/styled'
+import { Button } from '@hedvig-ui'
 import { PaperTitle } from 'components/claims/claim-details/components/claim-items/PaperTitle'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import { Button } from 'hedvig-ui/button'
 import { CardContent } from 'hedvig-ui/card'
 import { List, ListItem } from 'hedvig-ui/list'
 import { Spacing } from 'hedvig-ui/spacing'
