+++ src/components/claims/claim-details/components/ClaimType.tsx
@@ -6,12 +6,12 @@ import {
   useSetClaimTypeMutation,
 } from 'api/generated/graphql'
 
+import { Button } from '@hedvig-ui'
 import {
   PaperTitle,
   PaperTitleBadgeProps,
 } from 'components/claims/claim-details/components/claim-items/PaperTitle'
 import { format, parseISO } from 'date-fns'
-import { Button } from 'hedvig-ui/button'
 import { CardContent } from 'hedvig-ui/card'
 import { DateTimePicker } from 'hedvig-ui/date-time-picker'
 import { Input } from 'hedvig-ui/input'
