+++ src/components/member/tabs/account-tab/BackfillSubscriptionsButton.tsx
@@ -1,5 +1,5 @@
+import { Button } from '@hedvig-ui'
 import { useBackfillSubscriptionsMutation } from 'api/generated/graphql'
-import { Button } from 'hedvig-ui/button'
 import React from 'react'
 import { toast } from 'react-hot-toast'
