+++ src/components/member/tabs/campaigns-tab/campaigns/CampaignsRedeemedTable.tsx
@@ -1,9 +1,9 @@
+import { Button } from '@hedvig-ui'
 import {
   RedeemedCampaign,
   useManualUnRedeemCampaignMutation,
 } from 'api/generated/graphql'
 import { format } from 'date-fns'
-import { Button } from 'hedvig-ui/button'
 import { Capitalized } from 'hedvig-ui/typography'
 import React from 'react'
 import { Table } from 'semantic-ui-react'
