+++ src/components/member/tabs/contracts-tab/agreement/InsuranceCertificate.tsx
@@ -1,9 +1,9 @@
+import { Button, ButtonLink, ButtonsGroup } from '@hedvig-ui'
 import { Contract, GenericAgreement } from 'api/generated/graphql'
 import {
   regenerateCertificateOptions,
   useRegenerateCertificate,
 } from 'graphql/use-regenerate-certificate'
-import { Button, ButtonLink, ButtonsGroup } from 'hedvig-ui/button'
 import { ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import Dropzone from 'react-dropzone'
