+++ src/components/member/tabs/contracts-tab/agreement/InsuranceMandate.tsx
@@ -1,5 +1,5 @@
+import { ButtonLink } from '@hedvig-ui'
 import { Contract } from 'api/generated/graphql'
-import { ButtonLink } from 'hedvig-ui/button'
 import { ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
