+++ src/components/member/tabs/contracts-tab/agreement/ToDate.tsx
@@ -1,10 +1,10 @@
+import { Button, ButtonsGroup } from '@hedvig-ui'
 import { Contract, GenericAgreement } from 'api/generated/graphql'
 import { format } from 'date-fns'
 import {
   changeToDateOptions,
   useChangeToDate,
 } from 'graphql/use-change-to-date-for-agreement'
-import { Button, ButtonsGroup } from 'hedvig-ui/button'
 import { DateTimePicker } from 'hedvig-ui/date-time-picker'
 import { Spacing } from 'hedvig-ui/spacing'
 import {
