+++ src/components/member/tabs/contracts-tab/contract/master-inception.tsx
@@ -1,10 +1,10 @@
+import { Button, ButtonsGroup } from '@hedvig-ui'
 import { Contract } from 'api/generated/graphql'
 import { format } from 'date-fns'
 import {
   activateContractOptions,
   useActivateContract,
 } from 'graphql/use-activate-contract'
-import { Button, ButtonsGroup } from 'hedvig-ui/button'
 import { DateTimePicker } from 'hedvig-ui/date-time-picker'
 import { FourthLevelHeadline, Paragraph } from 'hedvig-ui/typography'
 import React from 'react'
