+++ src/components/member/tabs/quote-tab/quote-list-item.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled'
+import { Button } from '@hedvig-ui'
 import { Contract, Quote } from 'api/generated/graphql'
 import { UpdateQuoteForm } from 'components/member/tabs/quote-tab/update-quote-form'
 import { format, parseISO } from 'date-fns'
-import { Button } from 'hedvig-ui/button'
 import { ErrorText, ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import { getSchemaDataInfo } from 'utils/quote'
