+++ src/components/shared/paginator/Paginator.tsx
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react'
 import styled from '@emotion/styled'
-import { Button, ButtonsGroup } from 'hedvig-ui/button'
+import { Button, ButtonsGroup } from '@hedvig-ui'
 import React from 'react'
 import { Table } from 'semantic-ui-react'
 import { range } from 'utils/helpers'
