+++ src/features/tools/employees/components/EmployeeTable.tsx
@@ -1,3 +1,4 @@
+import { Button, ButtonsGroup } from '@hedvig-ui'
 import {
   Employee,
   EmployeesDocument,
@@ -10,7 +11,6 @@ import {
   LoadingMessage,
   StandaloneMessage,
 } from 'hedvig-ui/animations/standalone-message'
-import { Button, ButtonsGroup } from 'hedvig-ui/button'
 import { Card } from 'hedvig-ui/card'
 import { Dropdown } from 'hedvig-ui/dropdown'
 import React, { useState } from 'react'
