+++ src/features/tools/staging-tools/unsign-member-tool/index.tsx
@@ -1,5 +1,5 @@
+import { Button } from '@hedvig-ui'
 import { useUnsignMemberMutation } from 'api/generated/graphql'
-import { Button } from 'hedvig-ui/button'
 import { Input } from 'hedvig-ui/input'
 import { Spacing } from 'hedvig-ui/spacing'
 import { MainHeadline } from 'hedvig-ui/typography'
